# web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssueUpdateIn": {
        "additionalProperties": false,
        "properties": {
          "clear_owner": {
            "default": false,
            "title": "Clear Owner",
            "type": "boolean"
          },
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          },
          "owner_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner User Id"
          },
          "status": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "IssueUpdateIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Relay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/issues/{issue_id}": {
      "get": {
        "operationId": "get_issue_api_v1_issues__issue_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Issue",
        "tags": [
          "workspace"
        ]
      },
      "patch": {
        "description": "Owner and workflow status. ``resolved`` is refused: only a current run resolves issues.",
        "operationId": "update_issue_api_v1_issues__issue_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Issue",
        "tags": [
          "issues"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "me_api_v1_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me",
        "tags": [
          "workspace"
        ]
      }
    },
    "/api/v1/migrations": {
      "get": {
        "description": "Portfolio: each migration's readiness on its latest run (stale when inputs changed).",
        "operationId": "list_migrations_api_v1_migrations_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MigrationSummaryOut"
                  },
                  "title": "Response List Migrations Api V1 Migrations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Migrations",
        "tags": [
          "workspace"
        ]
      },
      "post": {
        "operationId": "create_migration_api_v1_migrations_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MigrationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MigrationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Migration",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/v1/migrations/{migration_id}": {
      "get": {
        "operationId": "get_migration_api_v1_migrations__migration_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MigrationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Migration",
        "tags": [
          "workspace"
        ]
      }
    },
...
    "/api/v1/migrations/{migration_id}/audit-events": {
      "get": {
        "operationId": "list_audit_events_api_v1_migrations__migration_id__audit_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditEventOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Events",
        "tags": [
          "workspace"
        ]
      }
    },
    "/api/v1/migrations/{migration_id}/audit-events/verify": {
      "get": {
        "operationId": "verify_audit_api_v1_migrations__migration_id__audit_events_verify_get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainVerificationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Audit",
        "tags": [
          "workspace"
        ]
      }
    },
...
    "/api/v1/migrations/{migration_id}/datasets": {
      "get": {
        "operationId": "list_datasets_api_v1_migrations__migration_id__datasets_get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetOut"
                  },
                  "title": "Response List Datasets Api V1 Migrations  Migration Id  Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets",
        "tags": [
          "workspace"
        ]
      },
      "post": {
        "operationId": "create_dataset_api_v1_migrations__migration_id__datasets_post",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Dataset",
        "tags": [
          "setup"
        ]
      }
    },
...
    "/api/v1/migrations/{migration_id}/entity-candidates": {
      "get": {
        "operationId": "list_candidates_api_v1_migrations__migration_id__entity_candidates_get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CandidateOut"
                  },
                  "title": "Response List Candidates Api V1 Migrations  Migration Id  Entity Candidates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Candidates",
        "tags": [
          "workspace"
        ]
      }
    },
...
  }
}